Simplify KeySignatureMaps
KeySignatureMap objects are replaced with KeySignature objects. Instead
of tracking which modifier (sharp, flat, natural) should apply to every
note, a KeySignature tracks which notes should be affected by which
modifier. This makes it possible to mathematically generate signatures
for every mode, and will make it easier to track accidentals in a
measure.